pyproject.toml
[project]
name = "fastapi-app"
version = "0.1.0"
description = "FastAPI template with MongoDB + Beanie"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]",
    "beanie>=1.25,<1.29.0",
    "motor",
    "pydantic-settings",
    "python-multipart",
    "uvicorn[standard]",
    "python-jose[cryptography]>=3.5.0",
    "argon2-cffi>=25.1.0",
    "passlib[argon2]>=1.7.4",
    "user-agents>=2.2.0",
    "pytz>=2025.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
